# contracts/openapi/taymex-v1/openapi.generated.yaml
openapi: 3.1.0
info:
  title: TAYMEX Foundation API
  version: 0.1.0
  description: Canonical F5 HTTP contract for foundation and security proof routes.
servers:
- url: /
paths:
  /api/health:
    get:
      operationId: healthLiveness
      x-taymex-auth: public
      x-taymex-data-classification: internal
      x-taymex-cache: no-store
      responses:
        '200':
          description: Process liveness
  /api/health/ready:
    get:
      operationId: healthReadiness
      x-taymex-auth: public
      x-taymex-data-classification: internal
      x-taymex-cache: no-store
      responses:
        '200':
          description: Dependency readiness
  /api/auth/sign-in:
    post:
      operationId: authSignIn
      x-taymex-auth: public
      x-taymex-data-classification: sensitive
      x-taymex-cache: no-store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: Authenticated session established
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionActorResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
  /api/auth/sign-out:
    post:
      operationId: authSignOut
      x-taymex-auth: session
      x-taymex-data-classification: sensitive
      x-taymex-cache: no-store
      responses:
        '204':
          description: Current session revoked
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /api/auth/session:
    get:
      operationId: authCurrentSession
      x-taymex-auth: session
      x-taymex-data-classification: sensitive
      x-taymex-cache: no-store
      responses:
        '200':
          description: Current authenticated actor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionActorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/roles:
    post:
      operationId: adminCreateRole
      x-taymex-auth: session
      x-taymex-permission: identity.roles.manage
      x-taymex-assurance: AAL2
      x-taymex-data-classification: confidential
      x-taymex-cache: no-store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/accounts/{accountId}/roles:
    put:
      operationId: adminAssignRoles
      x-taymex-auth: session
      x-taymex-permission: identity.roles.manage
      x-taymex-assurance: AAL2
      x-taymex-data-classification: confidential
      x-taymex-cache: no-store
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRolesRequest'
      responses:
        '200':
          description: Account role set replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleAssignmentResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/audit/records:
    get:
      operationId: adminQueryAudit
      x-taymex-auth: session
      x-taymex-permission: audit.records.read
      x-taymex-assurance: AAL2
      x-taymex-data-classification: sensitive
      x-taymex-cache: no-store
      parameters:
      - name: actionCode
        in: query
        schema:
          type: string
          maxLength: 160
      - name: category
        in: query
        schema:
          type: string
          enum:
          - security
          - settings
          - administration
          - domain
          - data-access
          - system
      - name: correlationId
        in: query
        schema:
          type: string
          maxLength: 128
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
      responses:
        '200':
          description: Bounded audit records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRecordListResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
  /api/admin/settings/{key}:
    get:
      operationId: adminGetSetting
      x-taymex-auth: session
      x-taymex-permission: settings.values.manage
      x-taymex-assurance: AAL2
      x-taymex-data-classification: confidential
      x-taymex-cache: no-store
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          maxLength: 240
      - name: projectRef
        in: query
        schema:
          type: string
          maxLength: 240
      responses:
        '200':
          description: Effective setting value with provenance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingEffectiveResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    put:
      operationId: adminWriteSetting
      x-taymex-auth: session
      x-taymex-permission: settings.values.manage
      x-taymex-assurance: AAL2
      x-taymex-data-classification: confidential
      x-taymex-cache: no-store
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
          maxLength: 240
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingWriteRequest'
      responses:
        '200':
          description: Persisted setting version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingWriteResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
components:
  responses:
    ErrorResponse:
      description: Safe machine-readable error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    SignInRequest:
      type: object
      additionalProperties: false
      required:
      - email
      - password
      properties:
        email:
          type: string
          minLength: 3
          maxLength: 320
        password:
          type: string
          minLength: 1
          maxLength: 128
        clientLabel:
          type: string
          maxLength: 120
    SessionActorResponse:
      type: object
      additionalProperties: false
      required:
      - accountId
      - assurance
      properties:
        accountId:
          type: string
          format: uuid
        assurance:
          type: string
          enum:
          - AAL1
          - AAL2
    CreateRoleRequest:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - permissions
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 120
        permissions:
          type: array
          maxItems: 64
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 160
    RoleResponse:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - permissions
      - version
      properties:
        id:
          type: string
        name:
          type: string
        permissions:
          type: array
          items:
            type: string
        version:
          type: integer
    AssignRolesRequest:
      type: object
      additionalProperties: false
      required:
      - roleIds
      - expectedVersion
      properties:
        roleIds:
          type: array
          maxItems: 64
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 128
        expectedVersion:
          type: integer
          minimum: 0
    RoleAssignmentResponse:
      type: object
      additionalProperties: false
      required:
      - version
      properties:
        version:
          type: integer
    AuditRecordListResponse:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditRecordResponse'
    AuditRecordResponse:
      type: object
      additionalProperties: false
      required:
      - id
      - occurredAt
      - actionCode
      - category
      - severity
      - actorKind
      - actorId
      properties:
        id:
          type: string
          format: uuid
        occurredAt:
          type: string
          format: date-time
        actionCode:
          type: string
        category:
          type: string
        severity:
          type: string
        actorKind:
          type: string
          enum:
          - account
          - system
        actorId:
          type: string
        correlationId:
          type: string
        resourceType:
          type: string
        resourceId:
          type: string
    SettingEffectiveResponse:
      type: object
      additionalProperties: false
      required:
      - key
      - value
      - source
      properties:
        key:
          type: string
        value:
          oneOf:
          - type: string
          - type: integer
          - type: number
          - type: boolean
        source:
          type: string
    SettingWriteRequest:
      type: object
      additionalProperties: false
      required:
      - scope
      - value
      - expectedVersion
      properties:
        scope:
          type: string
          enum:
          - platform
          - project
        scopeRef:
          type: string
          maxLength: 240
        value:
          oneOf:
          - type: string
          - type: integer
          - type: number
          - type: boolean
        expectedVersion:
          type: integer
          minimum: 0
        source:
          type: string
          maxLength: 128
    SettingWriteResponse:
      type: object
      additionalProperties: false
      required:
      - key
      - scope
      - version
      properties:
        key:
          type: string
        scope:
          type: string
        scopeRef:
          type: string
        version:
          type: integer
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
      - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
          - code
          - category
          - messageKey
          - correlationId
          properties:
            code:
              type: string
            category:
              type: string
            messageKey:
              type: string
            field:
              type: string
            correlationId:
              type: string
x-engineering-platform-source-sha256: 16192eed173a7e406e63a0fcd0a208544be5f259298114d915a5a872573760ff